refactor(solver): name infer match expansion state (#14346)
